Separating out the management server settings handler from the chrome policy handler, since they will become increasingly different as we configure more settings that have nothing to do with chrome policy.
