+++ ufo/static/settingsConfiguration.html
@@ -3,7 +3,7 @@
 <link rel="import" href="bower_components/paper-button/paper-button.html" />
 <link rel="import" href="bower_components/paper-spinner/paper-spinner.html" />
 
-<dom-module id="chrome-policy-configuration">
+<dom-module id="settings-configuration">
   <style is="custom-style">
     #policy-config-form {
       margin-left: 40px;
@@ -38,7 +38,7 @@
   </template>
   <script>
     Polymer({
-      is: 'chrome-policy-configuration',
+      is: 'settings-configuration',
       properties: {
         resources: {
           type: Object,
