+++ ufo/templates/setup.html
@@ -29,10 +29,10 @@
     </header-container>
   </paper-display-template>
 
-  <paper-display-template resources="{{policy_config_resources}}">
-    <header-container resources="{{policy_config_resources}}">
-      <chrome-policy-configuration resources="{{policy_config_resources}}">
-      </chrome-policy-configuration>
+  <paper-display-template resources="{{settings_resources}}">
+    <header-container resources="{{settings_resources}}">
+      <settings-configuration resources="{{settings_resources}}">
+      </settings-configuration>
     </header-container>
   </paper-display-template>
